+++ Changelog.md
@@ -4,12 +4,34 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)//
 
-## [Unreleased: 3.1.0-rc.6]
+## [Unreleased: 3.2.0-rc.1]
 ### Fixed
+ - FIO-5731: fixed validation error for wizard forms with advanced conditions
+ - FIO-6966: Fixes empty address component data in CSV
  - FIO-6601: Fix getting submissions by reference from custom submissions collection
  - FIO-4500: Fixes an issue where specified in the component settings delimiter is not used for CSVExport of Tags component data
  - FIO-6414: fixed filtering issues inside DT
  - FIO-4216: Fixes 'Resource not found' issue when using submission collection and trying to retrieve submission using x-token
+ - FIO-6840: Refactor authentication to use case-insensitive query and fallback to $regex
+ - FIO-6729: changed the output of data in the logs
+ - FIO-3840: fixed swagger spec definition for components with obj type
+ - FIO-6406: fixed filter for Select inside DT
+ - FIO-4809: Added tests for Wizard suffix/prefix
+ - Bump json5 from 2.2.1 to 2.2.3
+ - Bump @xmldom/xmldom from 0.7.5 to 0.7.11
+ - Bump nunjucks from 3.2.3 to 3.2.4
+ - Bump vm2 from 3.9.11 to 3.9.18
+ - Bump semver from 7.3.8 to 7.5.2
+ - Bump cookiejar from 2.1.3 to 2.1.4
+ - FIO-6840: change to allow case insensitive exists endpoint
+ - FIO-7161: fixed issue with possible undefined value for radio component in csv export
+ - FIO-7067: fixed issue with empty value for patch request
+ - FIO-7166: Adding a mongodb feature check to core
+ - FIO-4216 | FIO-6601 - Fixes no custom submission collection lookup when getting submissions
+ - FIO-6919: fixed value for Signature submission in Data Tab
+
+### Changed
+ - FIO-6769: Removed the `empty` template
 
 ## 3.1.0-rc.5
 ### Fixed